chore(visualization): move all the visualization types from flows to a shared folder
