+++ src/flows/pipes/TestFlux/view.tsx
@@ -21,7 +21,7 @@ import {ViewType, RemoteDataState} from 'src/types'
 import {AppSettingContext} from 'src/flows/context/app'
 import {PipeContext} from 'src/flows/context/pipe'
 
-import {TYPE_DEFINITIONS} from 'src/flows/pipes/Visualization'
+import {TYPE_DEFINITIONS} from 'src/shared/visualization'
 import {_transform} from 'src/flows/pipes/Visualization/view'
 
 const TestFlux: FC<PipeProp> = ({Context}) => {
