+++ src/flows/pipes/Visualization/index.ts
@@ -1,20 +1,7 @@
-import {VisTypeRegistration} from 'src/types'
 import View from './view'
 import './style.scss'
 
-interface VisTypeRegistrationMap {
-  [key: string]: VisTypeRegistration
-}
-
-export const TYPE_DEFINITIONS: VisTypeRegistrationMap = {}
-
-const context = require.context('./types', true, /index\.(ts|tsx)$/)
-context.keys().forEach(key => {
-  const module = context(key)
-  module.default((def: VisTypeRegistration) => {
-    TYPE_DEFINITIONS[def.type] = def
-  })
-})
+import {TYPE_DEFINITIONS} from 'src/shared/visualization'
 
 export default register => {
   register({
