+++ src/shared/visualization/types/Band/index.tsx
