+++ src/shared/visualization/types/Band/properties.ts
