+++ src/shared/visualization/types/Gauge/GaugeOptions.tsx
