+++ src/shared/visualization/types/Gauge/ThresholdsSettings.tsx
