+++ src/shared/visualization/types/Gauge/index.tsx
