+++ src/shared/visualization/types/Gauge/properties.ts
