+++ src/shared/visualization/types/Graph/icon.tsx
