+++ src/shared/visualization/types/Heatmap/HeatmapOptions.tsx
