+++ src/shared/visualization/types/Heatmap/icon.tsx
