+++ src/shared/visualization/types/Heatmap/index.tsx
