+++ src/shared/visualization/types/Histogram/HistogramOptions.tsx
