+++ src/shared/visualization/types/Histogram/index.tsx
