+++ src/shared/visualization/types/Map/index.tsx
