+++ src/shared/visualization/types/Map/properties.ts
