+++ src/shared/visualization/types/Mosaic/icon.tsx
