+++ src/shared/visualization/types/Mosaic/properties.ts
