+++ src/shared/visualization/types/Scatter/ScatterOptions.tsx
