+++ src/shared/visualization/types/SingleStat/icon.tsx
