+++ src/shared/visualization/types/SingleStat/index.tsx
