+++ src/shared/visualization/types/SingleStatWithLine/SingleStatWithLineOptions.tsx
