+++ src/shared/visualization/types/SingleStatWithLine/icon.tsx
