+++ src/shared/visualization/types/Table/icon.tsx
