+++ src/shared/visualization/types/Table/properties.ts
